requirements.txt
fastapi
uvicorn[standard]
openai
mysql-connector-python
azure-cosmos
python-dotenv==1.0.1
pydantic
python-multipart